Standard methods moved to Importer (F1, Q1 parent)
save_consumption_if_needed
convert_string_to_datetime
