+++ orakwlum/importer/Q1Importer.py
@@ -3,7 +3,6 @@ __author__ = 'XaviTorello'
 # __name__ = "F1_importer"
 
 import logging
-from datetime import datetime
 
 from switching.input.messages import Q1, defs, PeriodeActiva
 from enerdata.profiles.profile import Profile
@@ -20,13 +19,13 @@ logger = logging.getLogger(__name__)
 class Q1Importer(Import):
     """
     Import F1 object
-        Use gisce/switching lib for parse F1 file
+        Use gisce/switching lib for parse Q1 file
         Use gisce/enerdata lib for profile and estimate
     """
 
     def __init__(self, file_to_import, collection="dataset"):
         """
-        Set F1 file and start parsing it
+        Set Q1 file and start parsing it
         """
 
         super(Q1Importer, self).__init__(file_to_import, collection)
@@ -36,7 +35,7 @@ class Q1Importer(Import):
 
     def parse(self):
         """
-        Parse a F1 object using gisce/switching F1
+        Parse a Q1 object using gisce/switching Q1
         """
         self.q1.set_xsd()
         self.q1.parse_xml()
@@ -48,23 +47,24 @@ class Q1Importer(Import):
     #   return "F1"
 
     @property
-    def invoices_count(self):
+    def lectures_count(self):
         return self.q1.num_factures
 
     @property
-    def invoices(self):
-        return self.q1.get_factures()
+    def lectures(self):
+        return self.q1.get_comptadors()
 
+    # todo adapt to Q1 parsed structure
     def process_consumptions(self):
         """
         Process all FacturaATR of the imported F1
         For each period create the related profile and estimate
         Create the related Consumption objecte and save it (if more significant) to DB
         """
-        invoices = self.invoices
+        lectures = self.lectures
 
-        # Get all FacturaATR invoices
-        for factura_atr in invoices['FacturaATR']:
+        # Get all FacturaATR lectures
+        for factura_atr in lectures['FacturaATR']:
             self.print_invoice_summary(factura_atr)
             periods, consumption_total = factura_atr.get_info_activa()
 
@@ -102,56 +102,6 @@ class Q1Importer(Import):
                     # Save (Upsert) consumtion following strategy "importance of data"
                     self.save_consumption_if_needed(consumption_from_measure)
 
-
-
-    # todo Take maths on Mongo to speed up it!
-    def save_consumption_if_needed(self, consumption_to_save):
-        """
-        Save a Consumption on datasource if needed.
-
-        Check the priority of importer data VS current data on datasource
-
-        If importer have more priority -> save it to datasource
-        """
-
-        importer_priority = consumption_to_save.SOURCE_PRIORITY[self.type]
-
-        # If not highest priority fetch the current data priority from datasource
-        if importer_priority > 0:
-            # Fetch current data from datasource and compare priorities
-            current_data = consumption_to_save.get_one(self.dataset, self.collection)
-
-            # If there aren't any data on datasource, ensure to save it the new one!
-            if not current_data:
-                current_data_priority = str(int(importer_priority) + 1)
-            else:
-                assert len(current_data) == 1, "There are more than one entries for this key...{}".format(current_data)
-
-                # todo on PROD erase if and enforce assert
-                #assert 'origin' in current_data[0], "No origin defined on datasource for this consumption\n'{}'".format(current_data)
-                # If not origin exists on source (older version), update this one
-                if 'origin_priority' not in current_data[0]:
-                    print "Origin not defined"
-                    current_data_priority = str(int(importer_priority) + 1)
-                else:
-                    current_data_priority = current_data[0]['origin_priority']
-
-        # Finally, IMPORTER vs DATA    ## if the same priority override datasource (rectifications)
-        if (importer_priority <= current_data_priority):
-            logger.info("Saving imported consumption to DB (priorities: '{} vs {}')".format(importer_priority, current_data_priority))
-            consumption_to_save.save(self.dataset, self.collection)
-
-
-
-
-    def convert_string_to_datetime(self, string):
-        """
-        Auxiliar tmp toDate function
-        """
-        # todo review what it's better
-        #   - assume 00 / 01h of each day  <----- now implemented
-        #   - fetch the <LecturaDesde><FechaHora> / <LecturaHasta><FechaHora> from F1 file
-        return datetime.strptime(string + " 0", "%Y-%m-%d %H")
 
 
     @staticmethod
